apps/cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "garmin-messenger-cli"
version = "0.1.0"
description = "CLI for Garmin Messenger (Hermes) — login, send, receive satellite messages"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "garmin-messenger",
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
garmin-messenger = "garmin_messenger_cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/garmin_messenger_cli"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
"**/garmin_messenger_cli/main.py" = ["E402"]
"**/tests/*" = ["N806"]

[tool.pytest.ini_options]
testpaths = ["tests"]
